Move TrillianClient out of pkg/util
pkg/util is required by pkg/client, while Rekor API
clients are not going to connect to Trillian directly.

Splitting this up removes 1.2 MB from a pkg/client user.
